requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0
django-heroku==0.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.5
sqlparse==0.4.2
whitenoise==5.3.0

djangorestframework~=3.13.1